Check resources are in expected states
Allow resource to check whether it is in an expected state and update
status and action accordingly.

This first step only allows the base Resource class to perform a check
that will always fail with a NotSupported error.  This allows for the
other resources to implement their own checks for their expected states.

Implements: blueprint stack-check (partial)
Change-Id: I18985476696faeaa47149c233a067f9494c2effa