Chef solo resource
Allow a compute resource to be bootstrapped with chef-solo. This
resource has two options for downloading cookbooks:

1). berkshelf/librarian-chef can be installed on the specified host,
allowing a Berksfile/Cheffile to be included with cookbook download
locations.

2). A "kitchen" can be specified which is a git url containing all
required cookbooks.

chef-solo is installed on the host via the Opscode omnibus installer,
the cookbooks are installed on the host, any databags, node files,
roles, etc... are created on the host, and chef-solo is called.

Change-Id: I951814b23a0a99927fea08fa02c1157525c7604a
Implements: blueprint chef-solo-resource